// src/shared/errors/code-errors/error-codes.enum.ts
export enum ErrorCodesEnum {
  BOS001 = 'BOS001',
  BOS002 = 'BOS002',
  BOS003 = 'BOS003',
  BOS004 = 'BOS004',
  BOS005 = 'BOS005',
  BOS006 = 'BOS006',
  BOS007 = 'BOS007',
  BOS008 = 'BOS008',
  BOS009 = 'BOS009',
  BOS010 = 'BOS010',
  BOS011 = 'BOS011',
  BOS012 = 'BOS012',
  BOS013 = 'BOS013',
  BOS014 = 'BOS014',
  BOS015 = 'BOS015',
  BOS016 = 'BOS016',
  BOS017 = 'BOS017',
  BOS018 = 'BOS018',
  BOS019 = 'BOS019',
  BOS020 = 'BOS020',
  BOS021 = 'BOS021',
  BOS022 = 'BOS022',
  BOS023 = 'BOS023',
  BOS024 = 'BOS024',
  BOS025 = 'BOS025',
  BOS026 = 'BOS026',
  BOS027 = 'BOS027',
  BOS028 = 'BOS028',
  BOS029 = 'BOS029',
  BOS030 = 'BOS030',
  BOS031 = 'BOS031',
  BOS032 = 'BOS032',
  BOS033 = 'BOS033',
  BOS034 = 'BOS034',
  BOS035 = 'BOS035',
  BOS036 = 'BOS036',
  BOS037 = 'BOS037',
  BOS038 = 'BOS038',
  BOS039 = 'BOS039',
  BOS040 = 'BOS040',
}